Starter code pipeline
I wrote a starter program to flesh out the basic structure of
the files. This program just generates a random initial configuration
of planets, then just has them fall in the -y direction. I used
graphics.py to animate the results. I also added a run.sh script
to run the entire pipeline (generate initial configuration, solve
it, and animate it). We can build on this going forward.